+++ changelog.md
@@ -2,3 +2,4 @@
 
 ## 0.1.2
 - Added missing `rootGetters` argument from the getters method [#4](https://github.com/jackmellis/mock-vuex/issues/4)
+- Creating a store with an action and mutation that share the same name did not work. The action would overwrite the mutation [#3](https://github.com/jackmellis/mock-vuex/issues/3)